backend/integrations/services.py
"""
Integration services for external systems
"""
from typing import Dict, Any
from decouple import config


class ServiceNowService:
    """Service for interacting with ServiceNow API"""

    def __init__(self):
        self.instance = config('SERVICENOW_INSTANCE', default=None)
        self.username = config('SERVICENOW_USERNAME', default=None)
        self.password = config('SERVICENOW_PASSWORD', default=None)

        if not all([self.instance, self.username, self.password]):
            raise ValueError(
                "ServiceNow credentials not configured. "
                "Please set SERVICENOW_INSTANCE, SERVICENOW_USERNAME, and SERVICENOW_PASSWORD in .env"
            )

    def create_change_request(
        self,
        short_description: str,
        description: str,
        priority: str,
        planned_start_date: str,
        planned_end_date: str
    ) -> Dict[str, Any]:
        """
        Create a change request in ServiceNow

        Args:
            short_description: Brief summary
            description: Detailed description
            priority: Priority level (1-4)
            planned_start_date: Start date (YYYY-MM-DD)
            planned_end_date: End date (YYYY-MM-DD)

        Returns:
            ServiceNow response with sys_id and number
        """
        # TODO: Implement actual ServiceNow API call
        # For now, this is a placeholder that would be implemented later
        # Example implementation:
        #
        # import requests
        # url = f"https://{self.instance}.service-now.com/api/now/table/change_request"
        # headers = {
        #     'Content-Type': 'application/json',
        #     'Accept': 'application/json'
        # }
        # data = {
        #     'short_description': short_description,
        #     'description': description,
        #     'priority': priority,
        #     'planned_start_date': planned_start_date,
        #     'planned_end_date': planned_end_date
        # }
        # response = requests.post(url, auth=(self.username, self.password), headers=headers, json=data)
        # response.raise_for_status()
        # return response.json()['result']

        raise NotImplementedError("ServiceNow API integration not yet implemented")

    def get_change_request(self, sys_id: str) -> Dict[str, Any]:
        """
        Get a change request from ServiceNow

        Args:
            sys_id: ServiceNow system ID

        Returns:
            ServiceNow change request data
        """
        raise NotImplementedError("ServiceNow API integration not yet implemented")

    def update_change_request(self, sys_id: str, **kwargs) -> Dict[str, Any]:
        """
        Update a change request in ServiceNow

        Args:
            sys_id: ServiceNow system ID
            **kwargs: Fields to update

        Returns:
            ServiceNow response
        """
        raise NotImplementedError("ServiceNow API integration not yet implemented")


class JiraService:
    """Service for interacting with Jira API"""

    def __init__(self):
        self.url = config('JIRA_URL', default=None)
        self.email = config('JIRA_EMAIL', default=None)
        self.api_token = config('JIRA_API_TOKEN', default=None)

        if not all([self.url, self.email, self.api_token]):
            raise ValueError(
                "Jira credentials not configured. "
                "Please set JIRA_URL, JIRA_EMAIL, and JIRA_API_TOKEN in .env"
            )

    def create_issue(self, project: str, summary: str, description: str) -> Dict[str, Any]:
        """
        Create a Jira issue

        Args:
            project: Project key
            summary: Issue summary
            description: Issue description

        Returns:
            Jira issue data
        """
        raise NotImplementedError("Jira API integration not yet implemented")


class GitHubService:
    """Service for interacting with GitHub API"""

    def __init__(self):
        self.token = config('GITHUB_TOKEN', default=None)

        if not self.token:
            raise ValueError(
                "GitHub token not configured. "
                "Please set GITHUB_TOKEN in .env"
            )

    def create_pull_request(
        self,
        repo: str,
        title: str,
        body: str,
        head: str,
        base: str = 'main'
    ) -> Dict[str, Any]:
        """
        Create a GitHub pull request

        Args:
            repo: Repository name (owner/repo)
            title: PR title
            body: PR description
            head: Branch to merge from
            base: Branch to merge into

        Returns:
            GitHub PR data
        """
        raise NotImplementedError("GitHub API integration not yet implemented")
